pc-server/pyproject.toml
[project]
name = "xbox-controller-server"
version = "0.1.0"
description = "Xbox Controller PC Server"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "vgamepad>=0.1.0",
    "websockets>=12.0",
    "qrcode[pil]>=7.4",
    "Pillow>=10.0",
    "pystray>=0.19.5",
    "netifaces>=0.11.0",
    "pynput>=1.7.6"
]
